Refactor check and collect to be flatter
