+++ pyproject.toml
@@ -52,7 +52,7 @@ line-length = 120
 [tool.flake8]
 doctests = true
 max-line-length = 120
-max-complexity = 48  # TODO: get this down to ~15 by replacing cli.check() with a class
+max-complexity = 15  # forces us to move code to separate function if it gets too complex
 extend-ignore = "E2,E3"  # may be incompatible to black, but also we use black anyway
 
 [tool.ruff]
